Fixed a bug where the scrollrange was calculated wrong
First the scrollrange would allow the first element to be
collapsed which doesn't make sense now that it's scrolling
off. It could also be wrong if the notification was a group.
Then the scrollrange was also wrong because the bottomInset
was always added even though it shouldn't have done that.

Change-Id: I7184946c4406bc0b5796194af86155aa6ba68c8e
Fixes: 28015188

Former-commit-id: a424c504d585fa948ec0bacb692af1e2260fc9cf