Fix opening archives in Downloads.
When opening an archive in Downloads we first want to use
the ACTION_MANAGE_DOCUMENT intent, as it's a file on Downloads.

However, for files within archives we don't want call this intent,
as ACTION_MANAGE_DOCUMENT does not support files in archives.

Finally, we actually need to call ACTION_MANAGE_DOCUMENT for
archive files on Downloads, to give third party apps a chance
to show up in the Intent choose (invoked by the Trampoline).

This CL makes things work as they worked before merging
DownloadsActivity with DocumentsActivity.

Bug: 28033554
Change-Id: I287759aa8fc0457341ce6ece07d5c2ccf36f2c8c

Former-commit-id: 8fa99701491dee0ac18b8707e32252cb3c621af1