Fixed bug with cropping out drop shadow for pinned stack.
The surface crop we apply to windows in the pinned stack was based on
the stack bounds. However, the windows in the pinned stack have a
drop shadow that extends outside the stack bounds to make them appear
floating so they were out displayed. We now extand the crop for windows
in stacks that have shadows and occupy the entire stack space to that
their drop shadows are visible.

Bug: 27364161
Change-Id: Ic2a94b091a93e7145a5455b494f0b689118eb5e3

Former-commit-id: 6bab4cf52d53074b7d353738ebd014c72366774d