Use DisplayNameWithDialect only for some locales
Although DisplayNameWithDialect seems to return friendlier, more
"casual" names (e.g. "American English"), the result was inconsistent,
and (at times) debatable. And since this setting affects not only
the language of the translation, but a locale, names like "Flemish"
kind of lost the whole "locale / location" idea.

So we revert to use DisplayName for all but a few selected locales
(that we verified are better with the "dialect" form).

Bug: 27704583
Change-Id: I587081da1293cccac3cdabcd188a9b8160c233ea

Former-commit-id: 1b2e7adc8821a6de2cfc2ce23b0c555c1bb7f1fa