PopupWindow. Don't use -1 width/height for calculations.
In findDropDownPosition it's entirely possible for width/height
to be -1. If this is so, and the popup is anchored to the right,
we could fail to see that it is offscreen (since we think the width
is -1), and so fail to do our position adjustment to move it to the
left. I think this was previously covered up by window manager bugs
with child windows that requested coordinates outside of their
parent frame. To fix this, we ontinue to pass the same value to
the window manager, but use the width/height we expect to receive
for local layout calculations.

Bug: 28085451
Change-Id: Ia04ca3fcd17ad8819615b5ff42f7923462ce4b42

Former-commit-id: f6e801da1a45b2458679e20f5e6061442a434e1b