Adding some new multi-endpoint stuff.
Adding getVideoStateFromCallType method to ImsCallProfile, which basically
just breaks out some of the existing logic in getVideoStateFromImsCallProfile.
This is used to translate the CALL_TYPE_* from an external call to a
video state (used when pulling the call).

Added a new ImsReasonInfo code for when multi-endpoint fails to configure
because the modem does not support it.

Bug: 27458894
Change-Id: I226e79005dccf3e8cae30e4d448543adbe59f922

Former-commit-id: 9eb78515dc7c3b7d52170d240a66cda8a7bbeb88