Crash early when requesting a keyboard shortcut helper...
with a null IResultReceiver.

We're seeing Dialer crashes in this code path but they happen in the
main Handler loop so we can't see where the request with the null
request is coming from. Crashing earlier will hopefully give us a
stack trace that we can use to diagnose the issue.

Bug: 27963013
Change-Id: I60e4ef2ef328fa69790bbcaa4f196c02f7443296

Former-commit-id: 53b854ad7de5c22e03383b0f6538a493f601b7c9