Always send configuration changes to activity.
We are not sending override configuration changes to activities if the change in
configuration is not significant enough (i.e did resources will not be reloaded
due to the change). This breaks apps that rely on getting onConfigurationChanged()
call whenever the config changes regardless of if resources will be reloaded.
We no deliver all configuration changes to the activity.

Bug: 28177873
Bug: 23904868
Change-Id: Icc979dabb8bc93ae8a976ef6074cd2559aedca67

Former-commit-id: f7dab1e53f4c8544a6129ff8785363baec1272dc