When restacking IM windows preserve relative order.
We were still having some troubles with attached windows which we
werent properly promoting here. Now we catch everything with
the mIsImWindow flag. Before it was safe to assume dialogs
were on top of the Input Method but with a whole range of child
windows we need to be more careful. WindowManagerService has
been very careful preparing the window list for us, so carry
out its wishes and preserve the original relative ordering of
input method windows when restacking.

Bug: 28235787
Change-Id: Ied96d83f77b82df81b3b0aa94cb61e9c22999d23

Former-commit-id: b976877a4b62a6030fe67796ab8a6a69b4cc041f