BatteryStatsHelper: Include screen on baseline power draw
Calculate the baseline power usage for the device when it is in suspend and idle.
The device is drawing POWER_CPU_IDLE power at its lowest power state.
The device is drawing POWER_CPU_IDLE + POWER_CPU_AWAKE power when a wakelock is held.
The device is drawing POWER_CPU_IDLE + POWER_CPU_AWAKE + POWER_SCREEN_ON when the screen is on.

Bug:27533512
Change-Id: Idcb587390bc8159fcbd6625cca4cb1aca19976d6

Former-commit-id: 76a267bc265340824a394b6b310289f37dae9437