Fix IME for status bar in docked mode
If the dock divider is visible, window manager raises the IME from the
app's layer on top of the divider. However if the IME was targeting the
status bar, it would also remove it from the status bar's layer and
move it atop the divider (but below the status bar).

To fix this, we now only perform the adjustment to the IME's layer if
that moves the IME up, but never down.

Change-Id: I1308f51b98fffee64a5075c49697f5bc177ea32e
Fixes: 28024606

Former-commit-id: ae712e5708aeb2579bcfa38794e840350a992d97