Touch region fixes for dialogs and overlapping tasks.
There were two errors. First the function assumes that there are only
overlapping tasks in the case of freeform, and so otherwise it is not
necessary to union the focused task back at the end (and we can just
subtract all the non focused tasks). This won't work though, as there
can be for example dialogs over the launcher from another task. We
need to always add back the focused task. Secondly, we were requiring a
task to have a top visible "main app window" e.g. TYPE_APPLICATION or
TYPE_APPLICATION_STARTING. For tasks whose only presence on screen
is a dialog this will skip them and we will ignore them one way
or another.

Bug: 26962445
Change-Id: Idd5c54082409e7c34d9fc7480908c88fb12b378b

Former-commit-id: 9e76d3292b34c8683b549cc1f56bfff288defb92