fix "am start -W" hang on activity start
Normally launch time counts from the point when the activity is
resumed, to when the first window is drawn. However the activity
could become visible before it is resumed, due to some other
activity in the same task being launched. In this case we still
need to report launch time to unblock
ActivityStarter.startActivityMayWait().

bug: 28333487
Change-Id: I445b602f495ed0c3cb669e2d81a137280317534f

Former-commit-id: afb776d5447e19565c9a826a554911decb9ed92a