Don't hide app windows due to not showing when locked when keyguard is hidden
This worked in pre-N because the only visible app was the app that can be
shown when the keyguard is hidden. That isn't the case in multi-window mode
where one of the apps can be shown when locked and the other doesn't have
the show when locked flag. Only hide the other app if the keyguard is shown.

Bug: 28368875
Change-Id: I5039098db74492fadf667fed24fc58448436681a

Former-commit-id: 23fd1db68d1c28c8edc4ece1a0a1aa36c278e980