Remove onUhandledInputEvent API.
This callback was never used by WebView and anyone that overrides it
is going to just be broken since we never actually call into it from
the method that WebView does call. Furthermore, the WebView team has
decided that the API has too many complications to ever implement,
the most notable of which is that MotionEvents tend to come in
streams that would need to be buffered in order to detect the event
before it can be determined whether they were handled or not which
means that any uses of the API are inevitably going to be janky
experiences.

Bug: 14279909
Change-Id: I068601ce947bccacabfe55b86b06005449b65bcf

Former-commit-id: 1ad3955a6008f5f442fa5eb0adf7c4e6ba2f17cd