Do not lose window grab when approaching the edge of the screen
Currently when a freeform window is being dragged and the pointer gets
too close to the screen (stack) edge, the dragging completes. This was
done in ag/770762 to to prevent a freeform window from being dragged
fully under a docked stack.

This CL provides a different solution to this problem: while dragging a window
make sure that a big enough portion of its caption bar is still visible.

Bug: 26055288
Change-Id: I5873bb1203a0e91ae1436b3e87220affb77194ff

Former-commit-id: a90a3b85a5cf7fa2f27473064b2965d3c98f61a0