Launch home activity in home stack if coming from ResolverActivity
ResolverActivity launches activities it resolves as the initial caller.
In the case of the home intent it can be sys-ui, but only the system
server is currently allow to launch things into the home stack, so the
resolved home activity is placed in non-home stack. This wasn’t a
visible problem in fullscreen mode as the user would notice since
everything is fullscreen, however it is very visible in multi-window mode.
We now allow home activities to be places in the home stack if it is coming
from the ResolverActivity.

Bug: 28487506
Change-Id: I68f81da68a207efab9ce911fa6661bd573f1e949

Former-commit-id: 1affbbc5932231d5a1fa4563ef0623b4eaaa9cdd