Fix a behavior change in "Show software keyboard".
This bug was a corner case of stopping lying about
Configuration#keyboard when "Show software keyboard" is turned on, which
was done by my CL [1] in Android N development cycle.

Previous implementation

  if (config.keyboard != Configuration.KEYBOARD_NOKEYS) {

had relied on the fact that the system was lying about config.keyboard,
which is no longer valid.  We need to change the behavior of
InputMethodService#onShowInputRequested() depending on
Settings.Secure.SHOW_IME_WITH_HARD_KEYBOARD like we did for
InputMethodService#onEvaluateInputViewShown() in [1] to minimize the
impact on stopping lying about Configuration#keyboard.

 [1]: Id4d332e3909590c68345e10e1f2e18650efb2eb7
      a9093436c30622235611a90f055a34883d731568 [formerly 7b739a802cb7d97460a0bf8a1e9774efe96fa552]

Bug: 28423439
Change-Id: I8a774cbf61ac706d8446be91b17bceee57a13656

Former-commit-id: d0d07974745dcc947bb498198166f5872c7ec4a1