Let the framework position a window when switching to freeform
When "freeform" button is tapped, the current Recents implementation
puts the window into the left top corner, and sets window size
to the size of the window snapshot as it is displayed in Recents. This
is confusing.

This patch lets the ActivityManager decide on the window bounds.
The first "freeformed" window is put at the center, and subsequent
windows are nicely cascaded.

Bug: 28316396
Bug: 28312983
Change-Id: I3e305df3ab1e1f715e1f4f94e0ed2a8a1bbd2f22

Former-commit-id: 1f6b709ca67813e04212dec6eb48485b94753dfd