Delay the launch of the persistent task.
Looking at ActivityStarter, it seems like there is code to show
recents if the docked application is started from the home stack
and that ends up looking janky for us because we immediately switch
to the application we want to launch.

To avoid this, we defer the call to launch the docked stack until
after we get notified of a task launch on the fullscreen application
stack and avoid showing recents.

Change-Id: I92ba8aca3c29112f8c9520145aa00f6b58a1d31c

Former-commit-id: 02ee885a3ba41491a64567761d7f90a5f3db98f4