Fixed a bug where notifications could become orphaned
When a child got set not be a heads up after it was already
removed (When for example it was clicked on shortly after
it got posted) then it got orphaned leading to bad bugs
like the next notification not being visible. This
could happen often with WhatsApp who set the head up flags
on the summary instead of the children.

Change-Id: I52c8f5d9bc50080e4ece530255236d36f998372e
Fixes: 28565942

Former-commit-id: 52941c5618d79c0fb7af655f47f558d956af28c8