Fixed a crash with min-priority notifications in a group
Those notifications could get into a group with higher priority
because they were autobundled or if the app said so.
This could lead to invisible views on the lockscreen and a crash
when this happened in an animated fashion.
The bug was triggering old code that wasn't needed anymore that
removed the viewstate.

Change-Id: Ic86c2f03d55f2e98e352cdbfbb3ca80d8e676765
Fixes: 28611831

Former-commit-id: 06416b9e576369a1349ee7b60af62fae489fc098