Fix AbstractStringBuilder gaining extra APIs.
AbstractStringBuilder showed up in the API as gaining methods originally
belonging to StringBuffer/StringBuilder (mostly append(*) and
insert(*)).

Ideally it should not be in the API at all, but it's traditionally been
there and we'll keep it for now.

@hide the affected methods and move the "implements Appendable"
definition to StringBuffer/StringBuilder (as methods implementing an
interface can't be @hidden).

Bug: 27582395
Change-Id: I0b85c2c84f3e52ffe7864728d083f9c6ea7304ef

Former-commit-id: 91f4aec3ccb4b2a2b66fbde095baa584b0862c11