Fixed a bug where the positioning on the lockscreen was wrong
Because of a circular dependancy of calculating the notification
heights and setting them to GONE, the views could be laid out in a
wrong fashion. Another issue was that ambient notifications would
still be taken into account as well as removed notifications.

This also lead to a bug where the removal of notifications could
lead to unnecessary add animations when swiping them away.

Change-Id: If9ad81237a520b14dc6851df3af52406d192a7a7
Fixes: 28441832

Former-commit-id: fa760d46369f1ee7baedd174cc8f973594c52970