ZygoteInit: warm up JCA providers during preload
This makes the time spent in the first call of an app to
SSLSocketFactory.getDefault() drop from ~240 ms to ~50 ms. In M
it was around ~6ms. This is due to the fact that, while instantiating
the default factory, all providers are initialized.

In order to obtain the timings above, I created an app calling
SSLSocketFactory.getDefault in onCreate and timed it
with System.currentTimeMillis() .

(cherry picked from commit 724458ade6a06b52bff2def9012e5b1ebca69f34 [formerly 6cb7b1c4765e9bc5175056826523dbd88426e9aa])

Bug: 28545496

Change-Id: Ic5aab3ece609d9fef06fee4ccb83d8371af075b2

Former-commit-id: b07595d1941a25b5f82c62f9b7099c3c77d06056