Fix issue #28728838 java.lang.ClassCastException: android.os.BinderProxy...
...cannot be cast to LocalBinder after a crash

Turns out, if you called removeProcessLocked() with an old ProcessRecord,
we could get in to a bad state -- if we had a new process with that same
name already running, we would kind-of blow it away as a process name,
but not as much else.  So the next time we wanted to run a new thing
in the process, we would start a new second process and the two could
get mixed together.

Change-Id: I05e0463bd55f57f5e07f0b804ed9fd8c4f98a5a5

Former-commit-id: 4502ce2d498e6173f69b5af88cb2b81b76de990f