docs: Removed note that GET_ACCOUNTS was deprecated
We've decided not to deprecate GET_ACCOUNTS yet, but per emails,
we don't want to announce that it now *isn't* deprecated (because we
don't want to encourage people to use this permission); instead,
we're just going to not mention it at all.

Also cleaned up the following section on file permissions while I had
it open.

See first comment for doc stage location.

bug: 28727522
Change-Id: I27d21d186bdbf9ce905f3fb6a29027343a6daaf5

Former-commit-id: cc80393d94e53bc92f70e2d306ad87a92ff83794