Keep notifications with active remote inputs
Fixes a bad interaction when the app cancels the notification
into which the user is typing. This happens for example when
the user initiates direct reply from the lock screen and the
app is currently open underneath the lockscreen.

Also cleans up the code that clears notification effects by
moving it into the appropriate place (without functional changes).

Fixes: 28745456
Change-Id: I69e076642e93930d41cfb06f2a48836aea34d92d

Former-commit-id: 18d099a8fb3eb4332ce7f4cd43b1f00035e16bb4