Fixed flicker when docking task from recents.
Regression introduced by ag/1044640 where we don't want recents to the
visible if it's stack isn't fullscreen or focused. However, it is
possible for switching the home stack to not-fullscreen to prolonged
while transitioning to split-screen mode from recents. We now check
for if the docked stack exists which is a better indicator we are in
split-screen mode regardless of if the stask is fullscreen.

Bug: 28882948
Bug: 28246419
Change-Id: I54c02cc1486e7b5bf6a20dd0706c9bfe3f40a953

Former-commit-id: 8c09c7dd7a8838383a94ef85974a8ff0841dd8c5