Fix TextView layout reuse when maxLines is set
When maxLines is set on StaticLayout and the text is wrapped, during
onMeasure TextView couldn’t recognize the wrap in the layout and tried
to reuse the layout which in turn caused width calculation problems.
This CL checks if maxLines is set, and if there is a text cut-off in
order to reuse existing layout.

Bug: 28468120
Change-Id: Ide43df8512a09112715067cbadf29cc64fd53247

Former-commit-id: f75775ea7ff25fc9039ce7994312b29419294902