Enforce background fallback to be non-translucent
To avoid screen garbage or wallpaper shining through, the background
we use there for filling the whole needs to

a) Be non-translucent in all cases except if the window is translucent.
b) Don't infer it from the theme as it might be unexpected. Don't define
it in the theme, and use black if it's not defined.

Bug: 28763785
Change-Id: Ifb724b5c5feea3ed70d68e5bd7615fa91185e743

Former-commit-id: 8e879f1fec1ca6c641f270b58447033ab4671851