Fix action mode animation on recreations - framework edition
Currently if an action mode is started in onCreate()
it will fade in. This isn't ideal though, especially
since Activities are recreated routinely with
multi-window and resizable Activities. In that instance
we fade it in on every recreate.

This CL fixes this in both the decor and toolbar action
modes to only fade in if the decor has been laid out.

BUG: 29036694

Change-Id: Iae985efcced170a0a4229124c1c132355c2aa71e

Former-commit-id: e65b3fbb63b4b6937fede2e4b889c0a51a881358