Fix issue #29006049: Add logging for implicit broadcasts
We now have stats on broadcasts.  We collect them over a day and
then reset, retaining the last days stats.  When a checkin happens,
we return either the last day or the current stats and then clear
them.  Not bothing to persist anything to storage, this data is
not that critical.

Change-Id: I1c3b331bcd03f79fa5e10575d9bc2ad7d9104f6f

Former-commit-id: bc02a3901dea52d236dd855722191155156cb856