Set fontScale in override config to be equal to global.
The default fontScale in EMPTY config is 1.0. This will result in
updateFrom always overriding the fontScale to 1.0 since it's non-zero.

We can't set it to 0.0 either as that will make the override config
not equal to EMPTY. The rest of the code doesn't work well with that.

So here we always set the override fontScale to be idential to global.

bug: 29113700
Change-Id: I1dbe507145b8190a9ae5e108948c833d4e037e1d

Former-commit-id: 92ab103cb666059d48eca9e79977cf559abdde0a