Reinflate primary action mode view after detach
When app is resized in multi-window mode we preserve the window along
with the decor view. If action mode view was shown before such config
change, its view is detached, but the mPrimaryActionModeView variable
is not cleared. So when action mode view is shown again after that it
thinks that its view is still alive.
This CL adds additional check to inflate new mPrimaryActionModeView
if it is not attached to any window.

Bug: 28971666
Change-Id: Ia8c24d4322be32e8b2c8967301beb3a4d889d501

Former-commit-id: 8d6ac26504ae3708732e0f5c0df02efb17937d9b