Restore setCropInTransaction HiddenForCrop behavior.
Prior to c/1106850 setCropInTransaction hid the surface
for crop width and height <= 0. That CL allowed
setting -1 for crop width and height which SF
determines as clearing the crop. Other portions of the code
depend on the old behavior though for negative values, so restore
the behavior of setCropInTransaction and use a new clearCrop
method for the new code.

Bug: 29276588
Change-Id: I728666009c362ff635c7ebfb3ef2e83428fb03fe

Former-commit-id: 4320d33a065173dff97ddb462c8a6a0a649a41e5