Don't update task to return to type for last task over home
Resizing docked stack to fullscreen size results in dismissal
of docked stack and moving of tasks to fullscreen stack.
When tasks, which were originally launched from home or recents,
are inserted on top of fullscreen stack, their mTaskToReturnTo
value was overwritten.
This CL doesn't allow to overwrite this field of task if there
is no next task in stack and it was launched from home or recents.

Bug: 29237751
Change-Id: If56cd3f68b6b400c2d5c945431371acfe6269682

Former-commit-id: 21893824c1d15a1a46cb75725ad951620569c3c6