Make sure task exists in WM before trying to set the resizing state
It is possible for a task to be removed from the WM and stack in AM
while resizing (e.g. app crashes). In this case we want to make sure
the task exist in WM before we try to set the dock resizing state vs.
depending on the fact that the task is in recent list in AM.

Bug: 29325623
Change-Id: Ib89a7bfe12fe1f2c4c6c56ac9e13b0052ea2b4ec

Former-commit-id: d6aee18cabe75cbd4e98bfa380dffb89d89e0efc