A11y: Fix notification bounds when top notification is scrolled
Previously, scrolled notifications would collapse into a stack (the top
stack). With the new design, that's no longer the case, so tall
notifications can now be scrolled out so much that their translation
becomes negative.

Unfortunatly, getBoundsOnScreen then clips the bounds to the parent,
which interferes with our adjustment to the bounds, so we first undo
the clipping.

Fixes: 29368064
Change-Id: I71da054ed1eb6d06868d56e18d0320f764ed86b3

Former-commit-id: 98dd7f13085d1535b53bd22b235865f3f869ce35