Pinner Service - Pin Camera app on unlock if requested by the config
Pin the default camera app on unlock and user switch if the config requests it.
Do not pin an app if there is no default camera app for the intent.

Only pin the app for the MediaStore.INTENT_ACTION_STILL_IMAGE_CAMERA
action, not the _SECURE one.  While there could be different apps chosed
for different intents, the belief is most people will choose only one
app.

Pin both apk and odex.  Not all apps will have an odex, so it is not a
fatal error if that isn't pinned.

Bug 28251566

Change-Id: I3387c23c20d2f75c8864e29bdfc312fef6276434

Former-commit-id: 7cb2f8b7eeef57a21b1b1403ca1c0f4f5f381c35