Work on issue #29328569: NPE in " com.google.android.configupdater"
Make the activity manager look at the "killed" state of a process
before deciding it is in a state where a new client can attach to
a content provider.  Hopefully this will get rid of some races
around attaching to providers and processes being killed.

Also add a little more information when killing a client of a
provider describing the oom adj of that provider's process.

Change-Id: I4efe5daf0a96e4663d5df88827ef401b1311de3f

Former-commit-id: 66ec21a817db000ffc9868eddcf0fcb1caf9eee7