UsbMidiDevice: Clean up terminology and fix comments
When working with physical MIDI devices, an *input* stream is used
for reading from *output* port of the device, and vice versa. Thus,
using "input" and "output" without specifying whether it's a stream
or a port is confusing.

Clarify names of counter variables, and fix a couple of comments
that were incorrect due to this confusion. No functional changes.

Change-Id: If561eaca4bade94e9296d2c703c9fcebc91296e2

Former-commit-id: aab4c1ad489a8b62186215d826293e9cbcbaeb85