Always report window move to client
ag/1037916 added logic to only report window movement to the client
if the window doesn't have a task or its stack is bounds animating.
This causes problems on the client side for window that don't fall
into this category as they will have the wrong information of where
they are no screen.

Bug: 29093176
Change-Id: I958174af430f2c4003a1c0a74956964d209c0e4a

Former-commit-id: 7774bd554c28e33b732e449d029137537eb917da