On phones, don't show the avatar icon when it is the only user.
The keyguard user switcher isn't enabled for layouts under 600dp width.
For these thinner devices, we don't show the multi-user switch view when
there is only one user.

In order to update this visibility when user changes occur, we
re-calculate the visibilities when the keyguard status bar view is
made visible again.

Bug: 28795981
Change-Id: I980a4c400adc1d9a6c0ce3465854284997e86596

Former-commit-id: 8ba446c9bad94e91cf83c579261298552d51d249