Fix stale transformations
In messaging style we sometimes transform an individual
message and sometimes we don't. This causes the ViewTransformationHelper
to sometimes add the container and sometimes the individual messages,
which prevented the right view from getting reset.

To fix this, we now reset all views that were once managed by
ViewTransformationHelper but no longer are.

Change-Id: I1512434fde5b4444657db1c69f9a7bd1ba96fe5f
Fixes: 29509682

Former-commit-id: 75524417ac507597b19abb2a91389044bec682ad