fix last modified time
For cluster packages [i.e. packages that are split among multiple APKs]
the codePath is the directory containing the APKs. This normally works
well since the directory is created newly on update. However, it breaks
down during OTA when the directory modification time may or may not
change.

Specifically for cluster packages, we determine the packages "last
modified time" as the date of the newwest APK added to the package.

Bug: 29575840
Bug: 29610181

Change-Id: Ia0a40cf6aaf4102f35e9319ebdb95074b1740c90

Former-commit-id: 8051a0a3958f8ff75ce36cd5b4eb09dfc95662b1