Fixed bug with continuously changing orientation while dreaming
If the keyguard is locked while the device is dreaming in landscape
mode, the display will be frozen. When this happens we want to continue
to use the orientation the keyguard forced since that will be the
orientation used once the display is no longer frozen vs. switching to
the orientation of the top activity.

Bug: 29542506
Change-Id: I30a8370357735dcdff336a562ca248aa9206ae29

Former-commit-id: 049b6d89277f67ae210fdc4428df80a306556cfb