Add margins to the system icons when the user avatar is gone.
This resolves a regression caused by situtionally removing the
avatar on phones when there is only one user. Normally, the avatar's
width acts as padding against the end. By removing the avatar, the
buffer is also removed. This patch re-adds 6dp of margins to space
the system icons properly on phones.

Change-Id: I7b07b02fb74146e17109a7aaf23a41bb681683f3
FIXES: 29646859

Former-commit-id: f79cd873f6013489f29c66e4d8ea08f9253a71aa