Fix AttributeCache
AttributeCache relied on the frequency of the garbage collector
to clean up resource references. The garbage collector seems to have
gotten faster and the cache essentially doesn't cache anything.

Switch to using an LRU cache that holds strong references to
a finite set of Resources. This should be very small, just to
keep the few working apps involved in app switches in memory.

Bug:29405575
Change-Id: I64b34742c820f1e7872061a574e6cf1a76887149

Former-commit-id: 6aec452f91868c42b9dee156e9d2e22d1a1e5566