Remove "starting apps" boot message.
The "starting apps" message is only shown for a short period of time
before we show the lock screen.  Recent boot animations now have a
spin down phase which is just long enough to result in us flashing
this message for a very short period of time, which looks janky.

To avoid the janky behavior, remove this message altogether.  Other
boot messages for dexopt'ing after an OTA remain intact.

Also fix security bug that would allow any app to show a message.

Bug: 29367890
Change-Id: I77bb65acbe0e5afb4033dc6b83e533e665dce690

Former-commit-id: 7380c321ad23b6cbf0af239fbd2e037f67b08545