Use the first Vulkan layer path
The loader will scan the current layer path once when the first valid
Vulkan command call is made.  To avoid confusions and data races, never
set the layer path more than once.

This assumes the first layer path to be the app's library search path,
which should be the case.

Bug: 29780602
Change-Id: Ibb9e4de95902ee89413d213d237ea4f62ea35b38

Former-commit-id: 15e74c9614888d5b6acabc620e42d4e9e09cd5b1