Show notification while PRE_BOOT receivers run.
Awhile back, we switched to running PRE_BOOT receivers after the user
is unlocked.  Since these receivers include providers like Contacts
and MediaStore that can take a long time to upgrade, it may result in
apps like Dialer or Photos appearing to hang, so we need to help the
user understand that the system is still upgrading in the background.

Bug: 28164677
Change-Id: I18448534b6067f8508b1ffb2d43c106c02c265a0

Former-commit-id: c4bcacb98e8e820f11bbe95c1447b3f360ee6d0d